Moving utility/test helper files into subfolder include/util and src/util
