+++ extras/test/src/util/OTATestData.cpp
@@ -6,7 +6,7 @@
    INCLUDE
  **************************************************************************************/
 
-#include <OTATestData.h>
+#include <util/OTATestData.h>
 
 #include <algorithm>
